Split up err component mods to prepare for 0.10.0
